step one: consolidate daily size and motion data from pupil tracking csv's
